refactor(menu): rename to kebab-case
