+++ packages/components/menu/__tests__/menu.spec.ts
@@ -4,9 +4,9 @@ import { sleep } from '@element-plus/test-utils'
 import { rAF } from '@element-plus/test-utils/tick'
 
 import Menu from '../src/menu'
-import MenuGroup from '../src/menuItemGroup.vue'
-import MenuItem from '../src/menuItem.vue'
-import SubMenu from '../src/submenu.vue'
+import MenuGroup from '../src/menu-item-group.vue'
+import MenuItem from '../src/menu-item.vue'
+import SubMenu from '../src/sub-menu.vue'
 
 const _mount = (template: string, options = {}) =>
   mount({
