+++ packages/components/menu/index.ts
@@ -1,9 +1,9 @@
 import { withInstall, withNoopInstall } from '@element-plus/utils/with-install'
 
 import Menu from './src/menu'
-import MenuItem from './src/menuItem.vue'
-import MenuItemGroup from './src/menuItemGroup.vue'
-import SubMenu from './src/submenu.vue'
+import MenuItem from './src/menu-item.vue'
+import MenuItemGroup from './src/menu-item-group.vue'
+import SubMenu from './src/sub-menu.vue'
 
 export const ElMenu = withInstall(Menu, {
   MenuItem,
