+++ packages/components/menu/src/menu.ts
@@ -13,7 +13,7 @@ import {
 import { Resize } from '@element-plus/directives'
 import Menubar from '@element-plus/utils/menu/menu-bar'
 import ElMenuCollapseTransition from './menu-collapse-transition.vue'
-import ElSubMenu from './submenu.vue'
+import ElSubMenu from './sub-menu.vue'
 import { useMenuCssVar } from './use-menu-css-var'
 import type { VNode, Ref, ComputedRef } from 'vue'
 import type {
